A SECURE FEDERATED DISTRIBUTED COMPUTING BASED FRAMEWORK FOR DATA INFORMATION MANAGEMENT
The Smart-Frame, a general structure for enormous information
data administration in keen lattices in light of distributed computing innovation. The essential thought is to set up distributed computing focuses at three progressive levels to oversee data: top, local and end-client levels. While each territorial cloud focus is responsible for handling what's more, overseeing provincial information, the top cloud level gives a worldwide perspective of the structure.Furthermore, keeping in mind the end goal to bolster security for the structure, have exhibited an answer based on personality based cryptography and character based intermediary re encryption. Accordingly, the proposed system accomplishes not just adaptability and adaptability additionally security highlights. In the proposed system, the system has actualized a proof-of-idea for our system with a straightforward character based administration for information privacy. The prompt next stride is to likewise bolster intermediary re-encryption for the system. The personality based proxy re encryption is used to identify the individuals and to provide a more secure transmission and reception of data through a distributed level of computing.